fix: Interpret compartment coordinates as being the compartment ends
The coordinates in .swc files are probably meant to be interpreted as the compartment centres. However, it is unclear to me how to resolve the ambiguity in compartment lengths in this case. For now, I set the soma to end at the origin (x=0,y=0,z=0) and start at (x=0,y=0,z=-soma.length), so theta=0. Three-compartment soma is interpreted as one cylindrical compartment for now.